% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.map.R
\name{as.map}
\alias{as.map}
\title{Convert a map from a 3 columns data.frame to a map object}
\usage{
as.map(x, order = T)
}
\arguments{
\item{x}{a 3 columns data.frame with marker names, chromosome, and position}

\item{order}{whether loci should be ordered by position in each chromosome. If FALSE, the order of original file is kept}
}
\description{
Convert a map from a 3 columns data.frame to a map object
}
\examples{
data(sorgho_p118)
genonmv2<-as.map(genomv2.df)
}
\author{
J.-F. Rami \email{rami@cirad.fr}
}

