% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplot.r
\name{maplot}
\alias{maplot}
\title{Title alternative to plot.map including mk names and position labelling}
\usage{
maplot(map, map2, locusname = TRUE, position = TRUE, interchr = 0,
  ymax = NA, cex = 0.6, decalcoef = 0.02, out = "screen", ech = 1,
  newdev = TRUE, maih = 1, xlim = 1, first = 0, filename = NULL,
  sw = rep(1, length(map)), locolors = NULL, boxw = 500, main = NULL,
  ...)
}
\arguments{
\item{...}{}
}
\description{
Title alternative to plot.map including mk names and position labelling
}

